Pushover refactor
